Replace oauth2client with google-auth
oauth2client is now deprecated, and the new lib to replace this is
`google-auth`.  This swaps functionality over to it and in the process
removes a bit of code.

This switches to using the public API in `google.auth.Credentials` for
management of credential lifetimes, with exception to keeping the mutex
during a `refresh()` operation. I didn't see anything obvious in the new
API that had that, and it felt like a good behavior to preserve.

Most tests passed without modification. There's one new conflict
for the `google` module between the GAE SDK provided packages and
google-auth package from pypi. The pytest setup now just injects the GAE
SDK path into the `google` module so those imports can resolve.

I tested this on a GAE deploy without any issues.